# src/gateway/api.yml
swagger: '2.0'
info:
  title: User Signup API
  version: 1.0.0
  description: API for user signup
host: localhost:3000
basePath: /api
schemes:
  - http
paths:
...
  /cart/add/{productId}:
    post:
      summary: Add product to cart
      description: Adds a product to the user's cart
      parameters:
        - name: productId
          in: path
          description: ID of the product to add to the cart
          required: true
          type: string
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        200:
          description: Product added to cart successfully
          schema:
            $ref: '#/definitions/Cart'
        404:
          description: Product not found
  /cart/decrement/{productId}:
    put:
      summary: Decrement item quantity in cart
      description: Decrements the quantity of a product in the user's cart
      parameters:
        - name: productId
          in: path
          description: ID of the product to decrement in the cart
          required: true
          type: string
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        200:
          description: Item quantity decremented successfully
          schema:
            $ref: '#/definitions/Cart'
        404:
          description: Product not found in cart
  /cart/remove-item/{productId}:
    delete:
      summary: Remove item from cart
      description: Removes a product from the user's cart
      parameters:
        - name: productId
          in: path
          description: ID of the product to remove from the cart
          required: true
          type: string
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        200:
          description: Item removed from cart successfully
          schema:
            $ref: '#/definitions/Cart'
        404:
          description: Product not found in cart
  /cart/clear:
    delete:
      summary: Clear user's cart
      description: Removes all items from the user's cart
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        200:
          description: Cart cleared successfully
          schema:
            type: object
            properties:
              message:
                type: string
                example: Cart cleared successfully
        404:
          description: Cart not found
  /cart/view:
    get:
      summary: View user's cart
      description: Retrieves the current user's cart
      tags:
        - Cart
      security:
        - BearerAuth: []
      responses:
        200:
          description: Cart retrieved successfully
          schema:
            $ref: '#/definitions/Cart'
        404:
          description: Cart not found
...
definitions:
  Cart:
    type: object
    properties:
      userId:
        type: string
      items:
        type: array
        items:
          $ref: '#/definitions/CartItem'
      total:
        type: number
      modifiedOn:
        type: string
  CartItem:
    type: object
    properties:
      productId:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: number
      imageUrl:
        type: string
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: Use 'Bearer <token>' format